requirements.txt
# Build tools
setuptools>=68.0.0
wheel>=0.40.0
pip>=25.2.0

# Flask app
Flask==2.3.2
requests==2.31.0

# ML / TensorFlow
numpy==1.24.3
tensorflow-cpu==2.13.1

# Production server
gunicorn

gdown==4.7.1
